S2EExecutor: onStateForkDecide signal incorrectly emitted
KLEE will call S2EExecutor::fork regardless of the branch's condition being a
symbolic expression or concrete value. It is not until
klee::Executor::concolicFork and klee::Executor::fork that the condition is
checked and a fork performed **only** if the condition is **not** a
klee::ConstantExpr.

However, the onStateForkDecide signal is emitted **before** this check for a
klee::ConstantExpr is actually made. This signal is emitted unconditionally,
which means that it will occur for every branch instruction in LLVM code,
regardless of whether it depends on a symbolic expression or not.

For example, this can cause the following problem. If the program counter is
checked in the onStateForkDecide signal handler, it may show addresses that
correspond to instructions that are not necessarily branch instructions. This
is because functions like __ldb_mmu (which get inserted into the generated
LLVM code when performing a memory read) also contain conditional branches that
have nothing to do with symbolic values. For example, one check is that
`*g_sqi.events.before_memory_access_signals_count != NULL`. This branch
condition has no symbolic expressions, however S2EExecutor::fork will still be
called and the onStateForkDecide signal will still be emitted. This may lead to
incorrect state fork decisions (e.g. if you are basing your decision on the
program counter).

To fix this problem we restrict the onStateForkDecide signal so that it is only
emitted when the branching condition is based on a non-constant symbolic
expression.