LinuxMonitors: refactored for proper PID tracking
There were major problems with the current implementation of `getPid`. In
summary, the approach used (reading the TSS to find the current_thread_info
object) was removed from versions 4.x of the kernel. In newer versions of the
kernel, the only way to get the current_task struct was via the exported
symbol.

We therefore added a new LinuxMonitor command (LINUX_INIT) that transmits this
and other information (i.e. the PID and TGID offsets in the task_struct for the
current kernel) to the LinuxMonitor plugin. This makes the plugin independent
of the underlying Linux kernel and architecture, and now it is up to the
modified S2E kernel to send this information to S2E.

This now allows all the plugins that rely on OSMonitor::getPid to actually work
correctly.