s2e-kvm-io: experimental feature to abort MMIO for VAPIC emulation
Requires VAPIC to be turned on in the KVM client.
For now, it's easier to just disable VAPIC as making it work is hard.
VAPIC emulation mutates the CPU state in the middle of the I/O handler,
it's tricky to properly restore state. Currently, the guest OS would
just hang.

The downside of turning off VAPIC is slower 32-bit Windows emulation
as this would trigger more KVM exits on APIC access when clearning
interrupts.

Signed-off-by: Vitaly Chipounov <vitaly@cyberhaven.io>