klee,s2e: add support for fork() without symbolic conditions
As discussed in [1] and [2], it would be useful if the user can
fork a state unconditionally in their plugin code and switch back to it later.

This commit introduces a new way to fork a state without specifying
symbolic conditions. Essentially, an overloaded version of
`klee::Executor::fork()` is added which works similarly to
the regular version but without any path constraint code.

Besides, an overloaded version of `S2EExecutor::fork()` is added,
so that when the user explicitly forks a state using the new API,
S2E can write debugging messages just like the regular version of
`S2EExecutor::fork()` does.

Finally, since the two versions of `S2EExecutor::fork()` shares
a lot of similarity, the actual forking work has been moved to
`S2EExecutor::doFork()` in order to avoid code duplication.

Reference:
[1] https://groups.google.com/g/s2e-dev/c/x8tDX8i_v5o/m/JmuWEA4rQTAJ
[2] https://groups.google.com/g/s2e-dev/c/SeBpMlyHx3A/m/QqeOUSguDnIJ

Signed-off-by: Marco Wang <m.aesophor@gmail.com>