libtcg: fix use after free bug in tcg_register_helper()
Adding an element using g_array_append_val() may cause re-allocation
of the underlying array to accommodate the inserted element. However,
helper_table stores the pointer to the element in the underlying
array of garray (i.e., TCGHelperInfo*) as the value. If the
underlying array is re-allocated, accessing through helper_table may
cause the dereference of a wild pointer and further the crash/dead
loop of S2E.

To fix this bug, we just allocate the memory for TCGHelperInfo* helper
ourselves to avoid using the pointer to the private data field of
garray. After the fix, s_additional_helpers will store pointers
instead of instances of TCGHelperInfo.