makefile,dockerfile: Support using binary Z3 distributions
Adds support for building S2E against a binary distribution of Z3
and makes it the default option.

The dockerfile is updated to install libgomp1 and unzip (to run and
install a binary Z3 distribution), and to download and install a
recent version of CMake, which was necessary to build LLVM 13 from
source, so that these changes could be tested.