klee: always build with EH on, even if LLVM has it off
Klee uses exceptions. Pre-built LLVM comes with exceptions disabled.
By default KLEE tries to match LLVM's EH setting, but will fail to
build with EH off.

This commit always builds with EH on. It should not be an issue since
even an LLVM with EH off links against a libstdc++ with EH enabled, so
exceptions within KLEE should work.

This is required to build against a binary LLVM distribution.